🚚 engine.h: Extract `Size`
